simplify flattenFill
